backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langchain-groq
langchain-openai
qdrant-client>=1.7.0
python-dotenv
sentence-transformers
torch==2.4.0
--extra-index-url https://download.pytorch.org/whl/cpu
slowapi
tenacity
pydantic
pydantic-settings
python-multipart
pypdf
python-docx
docx2txt
tavily-python
httpx
beautifulsoup4
unstructured
supabase
upstash-redis
apscheduler
